# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateUserRequest:
    properties:
      email:
        description: Email is the user's email address.
        example: john.doe@example.com
        type: string
      first_name:
        description: FirstName is the user's first name.
        example: John
        maxLength: 100
        minLength: 2
        type: string
      last_name:
        description: LastName is the user's last name.
        example: Doe
        maxLength: 100
        minLength: 2
        type: string
      password:
        description: Password is the user's password.
        example: securepassword123
        minLength: 8
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.UserRole'
        description: Role is the user's role in the system.
        enum:
        - TEAM_MEMBER
        - PROJECT_MANAGER
        - ADMIN
        example: TEAM_MEMBER
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  dto.ErrorResponse:
    properties:
      details: {}
      message:
        example: An error occurred
        type: string
    type: object
  dto.GenericSuccessResponse:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        description: Email is the user's email address.
        example: john.doe@example.com
        type: string
      password:
        description: Password is the user's password.
        example: securepassword123
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
...
  dto.TokenResponse:
    properties:
      token:
        example: random-token
        type: string
    type: object
...
  dto.UpdateUserRequest:
    properties:
      first_name:
        description: FirstName is the optional new first name of the user.
        example: Johnny
        maxLength: 100
        minLength: 2
        type: string
      last_name:
        description: LastName is the optional new last name of the user.
        example: Smith
        maxLength: 100
        minLength: 2
        type: string
    type: object
  dto.UserResponse:
    properties:
      current_project_id:
        description: CurrentProjectID is the optional ID of the user's current project.
        example: 1
        type: integer
      current_project_name:
        description: CurrentProjectName is the optional name of the user's current
          project.
        example: Website Redesign
        type: string
      email:
        description: Email is the user's email address.
        example: john.doe@example.com
        type: string
      first_name:
        description: FirstName is the user's first name.
        example: John
        type: string
      id:
        description: ID is the unique identifier of the user.
        example: 42
        type: integer
      last_name:
        description: LastName is the user's last name.
        example: Doe
        type: string
      role:
        description: Role is the user's role in the system.
        example: TEAM_MEMBER
        type: string
    type: object
  dto.UserSliceSuccessResponse:
    properties:
      count:
        example: 5
        type: integer
      data:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      message:
        example: Items found successfully
        type: string
    type: object
  dto.UserSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.UserResponse'
      message:
        example: Operation successful
        type: string
    type: object
...
  models.UserRole:
    enum:
    - ADMIN
    - PROJECT_MANAGER
    - TEAM_MEMBER
    type: string
    x-enum-varnames:
    - Admin
    - ProjectManager
    - TeamMember
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a token
      parameters:
      - description: Login credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Login successful
          schema:
            $ref: '#/definitions/dto.TokenResponse'
        "400":
          description: Bad request - Invalid credentials or input
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: User login
      tags:
      - Users
  /me:
    get:
      description: Retrieves details of the authenticated user
      produces:
      - application/json
      responses:
        "302":
          description: User found
          schema:
            $ref: '#/definitions/dto.UserSuccessResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Users
...
  /users:
    get:
      description: Retrieves a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: Users found
          schema:
            $ref: '#/definitions/dto.UserSliceSuccessResponse'
        "400":
          description: Bad request - Database error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Creates a new user with the provided details
      parameters:
      - description: User creation request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/dto.UserSuccessResponse'
        "400":
          description: Bad request - Invalid input or validation failure
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new user
      tags:
      - Users
  /users/{userId}:
    delete:
      description: Deletes a user based on the provided ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted
          schema:
            $ref: '#/definitions/dto.GenericSuccessResponse'
        "400":
          description: Bad request - Invalid ID or user not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a user
      tags:
      - Users
    get:
      description: Retrieves a user based on the provided ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "302":
          description: User found
          schema:
            $ref: '#/definitions/dto.UserSuccessResponse'
        "400":
          description: Bad request - Invalid ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Updates an existing user's details
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: User update request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "202":
          description: User updated
          schema:
            $ref: '#/definitions/dto.UserSuccessResponse'
        "400":
          description: Bad request - Invalid input or ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a user
      tags:
      - Users
...
schemes:
- http
- https
swagger: "2.0"